Extract more functions to utils.js; Move constants into constants.js
