+++ constants.js
@@ -1,0 +1,7 @@
+export const apiHost = 'https://ipfs.exokit.org/ipfs';
+export const packagesEndpoint = 'https://packages.exokit.org';
+export const tokenHost = 'https://tokens.webaverse.com';
+export const network = 'rinkeby';
+export const infuraApiKey = '4fb939301ec543a0969f3019d74f80c2';
+export const rpcUrl = `https://${network}.infura.io/v3/${infuraApiKey}`;
+export const port = 9999;
